% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-RedisContainer.R
\name{configWorkerContainerEnv,RedisParamContainer-method}
\alias{configWorkerContainerEnv,RedisParamContainer-method}
\title{Configure the RedisParam worker container}
\usage{
\S4method{configWorkerContainerEnv}{RedisParamContainer}(container, cluster, workerNumber, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{RedisParamContainer} object
}
\description{
Configure the RedisParam worker container
}
