% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-RedisContainer.R
\name{deregisterParallelBackend,RedisParamContainer-method}
\alias{deregisterParallelBackend,RedisParamContainer-method}
\title{Deregister the BiocParallel RedisParam backend}
\usage{
\S4method{deregisterParallelBackend}{RedisParamContainer}(container, cluster, verbose = FALSE, ...)
}
\arguments{
\item{container}{The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}

\item{...}{Not used}
}
\value{
No return value
}
\description{
Deregister the BiocParallel RedisParam backend. The \code{BiocParallel::SerialParam()} will
be registered
}
