% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{RedisParamContainer-class}
\alias{RedisParamContainer-class}
\alias{.RedisParamContainer}
\title{The BiocParallel RedisParam worker container}
\description{
The BiocParallel RedisParam worker container
}

