% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-RedisContainer.R
\name{registerParallelBackend,RedisParamContainer-method}
\alias{registerParallelBackend,RedisParamContainer-method}
\title{Register the BiocParallel RedisParam backend}
\usage{
\S4method{registerParallelBackend}{RedisParamContainer}(container, cluster, verbose = FALSE, ...)
}
\arguments{
\item{container}{The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}

\item{...}{The additional parameter that will be passed to \code{RedisParam::RedisParam}}
}
\value{
No return value
}
\description{
Register the BiocParallel RedisParam backend. The registration will be done via
\code{RedisParam::RedisParam}
}
